+++ src/marshmallow/experimental/__init__.py
@@ -1,0 +1,5 @@
+"""Experimental features.
+
+The features in this subpackage are experimental. Breaking changes may be
+introduced in minor marshmallow versions.
+"""
